Mean snips without noise
Added  function tio calculate mean snips (for sipper cues and lick runs) that removed noise trials. It works fine but started to realise that a dictionary for each side might be best way to go - will be far easier to assign to actual substance and use in further bits of code.